FIWARE as source

+++ rivm-portaal-data.js
@@ -111,7 +111,7 @@
 //						"PM10":Influx.FieldType.INTEGER,
 						"PM10-eenheid": Influx.FieldType.STRING,
 						"PM10-meetopstelling": Influx.FieldType.STRING,
-						"PM":Influx.FieldType.INTEGER,
+						"PM":Influx.FieldType.FLOAT,
 //						"PM2.5":Influx.FieldType.INTEGER,
 						"PM-eenheid": Influx.FieldType.STRING,
 						"PM-meetopstelling": Influx.FieldType.STRING,
